# perf(core): remove promise ring